Use long for partitionVersion
This PR is to support the usecase where partition version
is calculated based on the last modified time of the
partition. For such cases, we need version to be long
since it depends on unix time.

+++ presto-hive-metastore/src/main/java/com/facebook/presto/hive/metastore/CachingHiveMetastore.java
@@ -618,7 +618,7 @@ public class CachingHiveMetastore
         HivePartitionName hivePartitionName = hivePartitionName(databaseName, tableName, partitionNameWithVersion.getPartitionName());
         Optional<Partition> partition = partitionCache.getIfPresent(hivePartitionName);
         if (partition != null && partition.isPresent()) {
-            Optional<Integer> partitionVersion = partition.get().getPartitionVersion();
+            Optional<Long> partitionVersion = partition.get().getPartitionVersion();
             if (!partitionVersion.isPresent() || partitionVersion.get() != partitionNameWithVersion.getPartitionVersion()) {
                 partitionCache.invalidate(hivePartitionName);
                 partitionStatisticsCache.invalidate(hivePartitionName);
